+++ flows/actions/base_test.go
@@ -211,7 +211,7 @@ func testActionType(t *testing.T, assetsJSON json.RawMessage, typeName string) {
 
 		// create an engine instance
 		eng := engine.NewBuilder().
-			WithWebhookServiceFactory(webhooks.NewServiceFactory(http.DefaultClient, "goflow-testing", 10000)).
+			WithWebhookServiceFactory(webhooks.NewServiceFactory(http.DefaultClient, map[string]string{"User-Agent": "goflow-testing"}, 10000)).
 			WithClassificationServiceFactory(func(s flows.Session, c *flows.Classifier) (flows.ClassificationService, error) {
 				if c.Type() == "wit" {
 					return wit.NewService(http.DefaultClient, c, "123456789"), nil