+++ flows/interfaces.go
@@ -114,6 +114,7 @@ type SessionAssets interface {
 	Classifiers() *ClassifierAssets
 	Fields() *FieldAssets
 	Flows() FlowAssets
+	Globals() *GlobalAssets
 	Groups() *GroupAssets
 	Labels() *LabelAssets
 	Locations() *LocationAssets